refactor: edit gatsby-config.js
