+++ gatsby-config.js
@@ -1,8 +1,13 @@
-const path = require("path")
-const remarkSlug = require("remark-slug")
 const siteMetadata = require("./site-metadata")
-const slashify = require("./src/utils/slashify")
-const { POSTS_PATH, IMAGES_PATH } = require("./src/utils/constants")
+const forceFileLoader = require("./gatsby/config/force-file-loader")
+const robotsTxt = require("./gatsby/config/robots-txt")
+const manifest = require("./gatsby/config/manifest")
+const svgrSvgo = require("./gatsby/config/svgr-svgo")
+const sharp = require("./gatsby/config/sharp")
+const mdx = require("./gatsby/config/mdx")
+const sourcePosts = require("./gatsby/config/source-posts")
+const sourceImages = require("./gatsby/config/source-images")
+const sitemap = require("./gatsby/config/sitemap")
 
 require("dotenv").config({
   path: `.env.${process.env.NODE_ENV}`,
@@ -24,142 +29,15 @@ module.exports = {
     siteUrl,
   },
   plugins: [
-    {
-      resolve: `gatsby-source-filesystem`,
-      options: {
-        name: `images`,
-        path: path.resolve(IMAGES_PATH),
-      },
-    },
-    {
-      resolve: `gatsby-source-filesystem`,
-      options: {
-        name: `posts`,
-        path: path.resolve(POSTS_PATH),
-      },
-    },
-    {
-      resolve: `gatsby-plugin-mdx`,
-      options: {
-        extensions: [`.mdx`, `.md`],
-        remarkPlugins: [remarkSlug],
-        gatsbyRemarkPlugins: [
-          {
-            resolve: `gatsby-remark-images`,
-            options: {
-              maxWidth: 900,
-              withWebp: true,
-              quality: 100,
-              loading: `lazy`,
-              linkImagesToOriginal: false,
-            },
-          },
-        ],
-      },
-    },
-    {
-      resolve: `gatsby-plugin-sharp`,
-      options: {
-        useMozJpeg: true,
-        defaultQuality: 100,
-      },
-    },
-    {
-      resolve: `gatsby-plugin-svgr-svgo`,
-      options: {
-        urlSvgOptions: [
-          {
-            test: /\.svg$/,
-            svgo: false,
-          },
-        ],
-      },
-    },
-    {
-      resolve: `gatsby-plugin-manifest`,
-      options: {
-        name: `Vladimir Skok`,
-        short_name: `skok.dev`,
-        description: `Software engineering blog`,
-        lang: `en`,
-        start_url: `/`,
-        background_color: `#fff`,
-        theme_color: `#11e`,
-        display: `standalone`,
-        icon: `static/favicon-dark.svg`,
-      },
-    },
-    {
-      resolve: `gatsby-plugin-sitemap`,
-      options: {
-        query: `{
-          allMdx {
-            nodes {
-              frontmatter {
-                datePublished
-                dateModified
-              }
-              fields {
-                slug
-              }
-            }
-          }
-        }`,
-        resolveSiteUrl: () => siteUrl,
-        resolvePages: ({ allMdx: { nodes: mdxNodes } }) => {
-          const { pages } = siteMetadata
-          const {
-            blog: { pathName: blogPathName },
-          } = pages
-
-          const allPages = Object.values(pages).reduce((acc, { pathName }) => {
-            if (pathName) {
-              acc.push({ path: slashify(pathName) })
-            }
-            return acc
-          }, [])
-
-          const allArticles = mdxNodes.map(
-            ({
-              frontmatter: { datePublished, dateModified },
-              fields: { slug },
-            }) => ({
-              path: slashify(blogPathName, slug),
-              lastmod: dateModified ? dateModified : datePublished,
-            })
-          )
-
-          return [...allPages, ...allArticles]
-        },
-        serialize: ({ path: url, lastmod }) => ({
-          url,
-          lastmod,
-        }),
-      },
-    },
-    {
-      resolve: `gatsby-plugin-robots-txt`,
-      options: {
-        resolveEnv: () => NETLIFY_ENV,
-        env: {
-          production: {
-            policy: [{ userAgent: `*` }],
-          },
-          "branch-deploy": {
-            policy: [{ userAgent: "*", disallow: ["/"] }],
-          },
-          "deploy-preview": {
-            policy: [{ userAgent: `*`, disallow: [`/`] }],
-          },
-        },
-      },
-    },
-    {
-      resolve: `gatsby-plugin-force-file-loader`,
-      options: {
-        rules: [`fonts`],
-      },
-    },
+    sourceImages,
+    sourcePosts,
+    mdx,
+    sharp,
+    svgrSvgo,
+    manifest,
+    sitemap,
+    robotsTxt,
+    forceFileLoader,
     `gatsby-plugin-image`,
     `gatsby-plugin-offline`,
     `gatsby-plugin-theme-ui`,
