+++ gatsby/config/force-file-loader.js
@@ -1,0 +1,8 @@
+const forceFileLoader = {
+  resolve: `gatsby-plugin-force-file-loader`,
+  options: {
+    rules: [`fonts`],
+  },
+}
+
+module.exports = forceFileLoader
