+++ gatsby/config/manifest.js
@@ -1,0 +1,16 @@
+const manifest = {
+  resolve: `gatsby-plugin-manifest`,
+  options: {
+    name: `Vladimir Skok`,
+    short_name: `skok.dev`,
+    description: `Software engineering blog`,
+    lang: `en`,
+    start_url: `/`,
+    background_color: `#fff`,
+    theme_color: `#11e`,
+    display: `standalone`,
+    icon: `static/favicon-dark.svg`,
+  },
+}
+
+module.exports = manifest
