+++ gatsby/config/sharp.js
@@ -1,0 +1,9 @@
+const sharp = {
+  resolve: `gatsby-plugin-sharp`,
+  options: {
+    useMozJpeg: true,
+    defaultQuality: 100,
+  },
+}
+
+module.exports = sharp
