+++ gatsby/config/svgr-svgo.js
@@ -1,0 +1,13 @@
+const svgrSvgo = {
+  resolve: `gatsby-plugin-svgr-svgo`,
+  options: {
+    urlSvgOptions: [
+      {
+        test: /\.svg$/,
+        svgo: false,
+      },
+    ],
+  },
+}
+
+module.exports = svgrSvgo
